Starts to add tests of the functionality
fuq was written as a prototype.  This is the first step toward making it solid.

The prototype has proven to be useful, and I think the code is in good enough shape that it doesn’t need to be thrown out and rewritten.  However, it does need some refactoring and some rearchitecting, but most importantly, it needs a suite of tests to make sure that things keep working.