Changes behavior of proto.Message As* methods
the AsOkay, AsError, and AsStop methods of proto.Message now no longer return an error and will panic() if the type is incorrect.

All of the existing uses check the type first, so this simplifies the interface quite a bit with no additional complexity.  The caller *should* check the type first, so the panic() is to report when the programmer doesn’t, which is an error.