Moves FetchJobs out of the JobQueuer interface
FetchJobs has rather complex logic, and currently works by scanning over all of the jobs.  It’s complex behavior to keep behind an interface.

For now, I’ve moved it into a separate function and added an EachJob method to the JobQueuer interface to iterate over all of the jobs.

At some point, if we need a more indexed-based way to query jobs, this may have to be revisited.