Fixes critical problem with autodialer
Foreman expects HELLO on reconnect, but worker nodes don’t know to issue this.  The Foreman now signals that it requires a HELLO by replying to an UPDATE with an OK(0xffff|0xffff).

If the dispatcher receives this, it will next send a HELLO message.  The current approach is a bit of a hack: Dispatch.sendUpdate returns a specific error that informs Dispatch.QueueLoop to send a HELLO.

We should replace this with something that’s less of a hack at some point, when we figure out what that is…